Extract URL utils into separate file (no grab bags here officer)
