+++ test/specs/utilities.test.js
@@ -5,42 +5,6 @@ const expect = chai.expect;
 
 const utilities = require('../../modules/utilities');
 
-const testRoutes = {
-    sections: {
-        purple: {
-            path: '/purple',
-            pages: {
-                monkey: {
-                    path: '/monkey/dishwasher',
-                    live: true,
-                    isWildcard: false,
-                    isPostable: false,
-                    aliases: ['/green/orangutan/fridge']
-                }
-            }
-        }
-    },
-    vanityRedirects: [
-        {
-            path: '/test'
-        },
-        {
-            paths: ['/supports', '/multiple', '/redirects']
-        }
-    ],
-    otherUrls: [
-        {
-            path: '/unicorns',
-            isPostable: true,
-            live: true
-        },
-        {
-            path: '/draft',
-            live: false
-        }
-    ]
-};
-
 describe('utility functions', () => {
     it('should strip trailing slashes correctly', done => {
         let pathWithSlash = '/foo/';
@@ -70,32 +34,6 @@ describe('utility functions', () => {
             expect(utilities.parseValueFromString(input)).to.equal(output);
         });
 
-        done();
-    });
-});
-
-describe('Cloudfront route generator', () => {
-    it('should filter out non-live routes', done => {
-        let urlList = utilities.generateUrlList(testRoutes);
-        expect(urlList.newSite.length).to.equal(9);
-        done();
-    });
-
-    it('should generate the correct section/page path', done => {
-        let urlList = utilities.generateUrlList(testRoutes);
-        expect(urlList.newSite.filter(r => r.path === '/purple/monkey/dishwasher').length).to.equal(1);
-        done();
-    });
-
-    it('should generate welsh versions of canonical routes', done => {
-        let urlList = utilities.generateUrlList(testRoutes);
-        expect(urlList.newSite.filter(r => r.path === '/welsh/purple/monkey/dishwasher').length).to.equal(1);
-        done();
-    });
-
-    it('should store properties against routes', done => {
-        let urlList = utilities.generateUrlList(testRoutes);
-        expect(urlList.newSite.filter(r => r.path === '/unicorns')[0].isPostable).to.equal(true);
         done();
     });
 });
